:truck: rename(Race): 패키지 변경
